requirements.txt
Django>=4.2
psycopg2>=2.9.9
djangorestframework
django-cors-headers
djangorestframework-simplejwt
